A variety of preparations for Apploop
- "inRenderPhase" bool for Windows
- matched FPS definitions with Timer stamping return values (as that is used for the game loop aswell)
- changed to c++ 20 (had to update visual studio trololol)
- tested semaphores in MyzticTest
- initMyztic now instantly redirects to a different thread so we can have the main thread accessible as a std::thread

TODO:
WinThread class to better handle threading instead of multiple std::map's in Application.h ??

+++ Myztic/cpp/display/Fps.cpp
@@ -5,7 +5,7 @@
 // Fps
 
 Fps::Fps(fpsSize max) {
-	frameTime = 1 / float(max);
+	frameTime = 1 / (float)max;
 	_used = max;
 	this->max = max;
 }
@@ -17,13 +17,13 @@ fpsSize Fps::getMax() {
 void Fps::setMax(fpsSize max) {
 	this->max = max;
 	_used = max;
-	frameTime = 1 / float(max);
+	frameTime = 1 / (float)max;
 }
 
 fpsSize Fps::used() {
-	return _used;
+	return _used; // This value is populated by the Application
 }
 
-float Fps::getFrameTime() {
+double Fps::getFrameTime() {
 	return frameTime;
 }
